% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnotgate.R
\name{CCNOT}
\alias{CCNOT}
\alias{toffoli}
\title{The CCNOT or toffoli gate}
\usage{
CCNOT(bits = c(1, 2, 3))

toffoli(bits = c(1, 2, 3))
}
\arguments{
\item{bits}{integer vector of length two, the first bit being the control and the second
the target bit.}
}
\value{
An S4 class 'ccnotgate' object is returned
}
\description{
The CCNOT or toffoli gate
}
