% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnotgate.R
\docType{class}
\name{ccnotgate}
\alias{ccnotgate}
\alias{ccnotgate-class}
\title{The CCNOT gate}
\description{
This class represents a generic CNOT gate
}
\section{Slots}{

\describe{
\item{\code{bits}}{Integer vector of length 2. First two bits are the control bits,
third the target bit.}
}}

\examples{
x <- qstate(nbits=3)
z <- CCNOT(c(1,2,3)) * (H(1) * x)

}
