% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnotgate.R
\name{CNOT}
\alias{CNOT}
\title{The CNOT gate}
\usage{
CNOT(bits = c(1, 2))
}
\arguments{
\item{bits}{integer vector of length two, the first bit being the control and the second
the target bit.}
}
\value{
An S4 class 'cnotgate' object is returned
}
\description{
The CNOT gate
}
