% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnotgate.R
\docType{class}
\name{cnotgate}
\alias{cnotgate}
\alias{cnotgate-class}
\title{The CNOT gate}
\description{
This class represents a generic CNOT gate
}
\section{Slots}{

\describe{
\item{\code{bits}}{Integer vector of length 2. First bit is the control bit,
second the target bit.}
}}

\examples{
x <- qstate(nbits=2)
## A Bell state
z <- CNOT(c(1,2)) * (H(1) * x)

}
