% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swapgate.R
\name{CSWAP}
\alias{CSWAP}
\alias{fredkin}
\title{The CSWAP or Fredkin gate}
\usage{
CSWAP(bits = c(1, 2, 3))

fredkin(bits = c(1, 2, 3))
}
\arguments{
\item{bits}{integer vector of length two, the first bit being the control and the second
the target bit.}
}
\value{
An S4 class 'cswapgate' object is returned
}
\description{
The CSWAP or Fredkin gate
}
