% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swapgate.R
\docType{class}
\name{cswapgate}
\alias{cswapgate}
\alias{cswapgate-class}
\title{The CSWAP gate}
\description{
This class represents a generic SWAP gate, also called Fredkin gate
}
\section{Slots}{

\describe{
\item{\code{bits}}{Integer vector of length 2. First two bits are the control bits,
third the target bit.}
}}

\examples{
x <- qstate(nbits=3)
z <- CSWAP(c(1,2,3)) * (H(1) * x)

}
