% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{genComputationalBasis}
\alias{genComputationalBasis}
\title{genComputationalBasis}
\usage{
genComputationalBasis(nbits, collapse = "")
}
\arguments{
\item{nbits}{integer. The number of qubits}

\item{collapse}{character. String to fill in between separate bits}
}
\value{
a character vector of length 2^nbits
}
\description{
function to generate the basis strings for given number of
bits
}
\examples{
genComputationalBasis(4)
genComputationalBasis(2, collapse=">|")

}
