% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{genNoise}
\alias{genNoise}
\title{genNoise}
\usage{
genNoise(nbits, p = 0, bits = 1:nbits, error = "any", ...)
}
\arguments{
\item{nbits}{integer. The number of qubits}

\item{p}{probability with which noise is applied after every gate}

\item{bits}{integer or integer array. The bit to which to apply the gate.}

\item{error}{String containing the error model.}

\item{...}{Additional arguments to be stored in \code{args}.}
}
\value{
a list containing \code{p}, \code{bits}, \code{error} and
\code{args}
}
\description{
function to generate the noise list
}
\details{
See function \code{noise} for details.
}
\examples{
genNoise(4)
genNoise(2, p=1, error="small", sigma=0.1)

}
