% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{genStateNumber}
\alias{genStateNumber}
\title{genStateNumber}
\usage{
genStateNumber(int, nbits)
}
\arguments{
\item{int}{integer number representing the basis state}

\item{nbits}{integer. The number of qubits}
}
\value{
a integer vector of length \code{nbits}
}
\description{
function to generate the bit representation for a specific basis state
}
\examples{
genStateNumber(5, 4)
genStateNumber(2, 2)

}
