% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{genStateString}
\alias{genStateString}
\title{genStateString}
\usage{
genStateString(int, nbits, collapse = "")
}
\arguments{
\item{int}{integer number representing the basis state}

\item{nbits}{integer. The number of qubits}

\item{collapse}{character. String to fill in between separate bits}
}
\value{
a character
}
\description{
function to generate the string for a specific basis state
}
\examples{
genStateString(5, 4)
genStateString(2, 2, collapse=">|")

}
