% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqgate.R
\name{H}
\alias{H}
\title{The Hadarmard gate}
\usage{
H(bit)
}
\arguments{
\item{bit}{integer. The bit to which to apply the gate}
}
\value{
An S4 class 'sqgate' object is returned
}
\description{
The Hadarmard gate
}
\examples{
x <- qstate(nbits=2)
z <- H(1) * x
z

}
