% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure.R
\name{hist.measurement}
\alias{hist.measurement}
\title{Plot the histogram of a quantum measurement}
\usage{
\method{hist}{measurement}(x, only.nonzero = TRUE, by.name = only.nonzero, freq = TRUE, ...)
}
\arguments{
\item{x}{object as returned by \code{measure}}

\item{only.nonzero}{are the states with zero measurements to be plotted?}

\item{by.name}{shall the xlabel contain the basis names? If \code{FALSE}, the
index number is used.}

\item{freq}{shall the total counts be plotted? If not, the values are
normalised to 1.}

\item{...}{Generic parameters to pass on to \code{barplot()}}
}
\value{
No return value.
}
\description{
Plot the histogram of a quantum measurement
}
