% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqgate.R
\name{Id}
\alias{Id}
\title{The identity gate}
\usage{
Id(bit)
}
\arguments{
\item{bit}{integer. The bit to which to apply the gate}
}
\value{
An S4 class 'sqgate' object is returned
}
\description{
The identity gate
}
\examples{
x <- qstate(nbits=2)
z <- Id(1) * x
z

}
