% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{is.bitset}
\alias{is.bitset}
\title{is.bitset}
\usage{
is.bitset(x, bit)
}
\arguments{
\item{x}{target vector}

\item{bit}{integer. The bit to check}
}
\value{
a boolean vector
}
\description{
checks whether or not a bit is set in target
}
