% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{normalise}
\alias{normalise}
\title{normalise}
\usage{
normalise(x)
}
\arguments{
\item{x}{complex valued vector}
}
\value{
Returns the normalised complex valued vector
}
\description{
Normalises a complex vector to 1
}
