% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-qstate.R
\name{plot,qstate,missing-method}
\alias{plot,qstate,missing-method}
\alias{plot}
\title{plot-qstate}
\usage{
\S4method{plot}{qstate,missing}(x, y, ...)
}
\arguments{
\item{x}{qstate object}

\item{y}{not used here}

\item{...}{additional parameters to be passed on}
}
\value{
nothing is returned, but a plot created
}
\description{
Plots a circuit corresponding to a qstate object
}
\examples{
x <- qstate(2)
y <- H(1) * x
z <- CNOT(c(1,2)) * y
plot(z)

}
