% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsimulatR-package.R
\docType{package}
\name{qsimulatR}
\alias{qsimulatR}
\title{The qsimulatR Package}
\description{
A simulator for a quantum computer
}
\details{
A quantum computer simulator framework. General single qubit gates
and general controlled single qubit gates can be easily defined.
For convenience, it currently directly provides
most common gates (X, Y, Z, H, Z, S, T, Rx, Ry, Rz, CNOT, SWAP,
toffoli or CCNOT, CSWAP). 'qsimulatR' supports plotting
of circuits and is able to export circuits into IBM's 'Qiskit' python
package, which can be run on IBM's real quantum hardware.
'qsimulatR' currently works for up to 24 qubits (a virtual
restriction, which can be lifted).
}
\author{
Johann Ostemeyer, Carsten Urbach, \email{urbach@hiskp.uni-bonn.de}
}
\keyword{package}
