% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqgate.R
\name{Rz}
\alias{Rz}
\title{The Rz gate}
\usage{
Rz(bit, theta = 0)
}
\arguments{
\item{bit}{integer. The bit to which to apply the gate}

\item{theta}{numeric. angle}
}
\value{
An S4 class 'sqgate' object is returned
}
\description{
The Rz gate
}
\examples{
x <- qstate(nbits=2)
z <- Rz(1, pi/4) * x
z

}
