% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqgate.R
\name{S}
\alias{S}
\title{The S gate}
\usage{
S(bit)
}
\arguments{
\item{bit}{integer. The bit to which to apply the gate}
}
\value{
An S4 class 'sqgate' object is returned
}
\description{
The S gate
}
\examples{
x <- X(1) * qstate(nbits=2)
z <- S(1) * x
z

}
