% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure.R
\name{summary.measurement}
\alias{summary.measurement}
\title{Summarize a quantum measurement}
\usage{
\method{summary}{measurement}(object, ...)
}
\arguments{
\item{object}{as returned by \code{measure}}

\item{...}{Generic parameters to pass on, not used here.}
}
\value{
No return value.
}
\description{
Summarize a quantum measurement
}
