% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swapgate.R
\name{SWAP}
\alias{SWAP}
\title{The SWAP gate}
\usage{
SWAP(bits = c(1, 2))
}
\arguments{
\item{bits}{integer vector of length two, containing the bits to swap.}
}
\value{
An S4 class 'swapgate' object is returned
}
\description{
The SWAP gate
}
