% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swapgate.R
\docType{class}
\name{swapgate}
\alias{swapgate}
\alias{swapgate-class}
\title{The SWAP gate}
\description{
This class represents a generic SWAP gate
}
\section{Slots}{

\describe{
\item{\code{bits}}{Integer vector of length 2. The two bits to swap.}
}}

\examples{
x <- H(1) * qstate(nbits=2)
z <- SWAP(c(1,2)) * (H(1) * x)

}
