% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqgate.R
\name{Tgate}
\alias{Tgate}
\title{The Tgate gate}
\usage{
Tgate(bit)
}
\arguments{
\item{bit}{integer. The bit to which to apply the gate}
}
\value{
An S4 class 'sqgate' object is returned
}
\description{
The Tgate gate
}
\examples{
x <- X(1)*qstate(nbits=2)
z <- Tgate(1) * x
z

}
