% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnotgate.R
\name{*,ccnotgate,qstate-method}
\alias{*,ccnotgate,qstate-method}
\title{times-ccnotgate-qstate}
\usage{
\S4method{*}{ccnotgate,qstate}(e1, e2)
}
\arguments{
\item{e1}{object of S4 class 'ccnotgate'}

\item{e2}{object of S4 class 'qstate'}
}
\value{
An object of S4 class 'qstate'
}
\description{
Applies a CCNOT (or toffoli) gate to a quantum state.
}
