% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccqgate.R
\name{*,ccqgate,qstate-method}
\alias{*,ccqgate,qstate-method}
\title{times-ccqgate-qstate}
\usage{
\S4method{*}{ccqgate,qstate}(e1, e2)
}
\arguments{
\item{e1}{object of S4 class 'ccqgate'}

\item{e2}{object of S4 class 'qstate'}
}
\value{
An object of S4 class 'qstate'
}
\description{
Applies a twice controlled single qubit gate to a quantum state.
}
