% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnotgate.R
\name{*,cnotgate,qstate-method}
\alias{*,cnotgate,qstate-method}
\title{times-cnotgate-qstate}
\usage{
\S4method{*}{cnotgate,qstate}(e1, e2)
}
\arguments{
\item{e1}{object of S4 class 'cnotgate'}

\item{e2}{object of S4 class 'qstate'}
}
\value{
An object of S4 class 'qstate'
}
\description{
Applies a CNOT gate to a quantum state.
}
