% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnqgate.R
\name{*,cnqgate,qstate-method}
\alias{*,cnqgate,qstate-method}
\title{times-cnqgate-qstate}
\usage{
\S4method{*}{cnqgate,qstate}(e1, e2)
}
\arguments{
\item{e1}{object of S4 class 'cnqgate'}

\item{e2}{object of S4 class 'qstate'}
}
\value{
An object of S4 class 'qstate'
}
\description{
Applies n-fold controlled single qubit gate to a quantum state.
}
