% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{*,complex,qstate-method}
\alias{*,complex,qstate-method}
\title{times-number-qstate}
\usage{
\S4method{*}{complex,qstate}(e1, e2)
}
\arguments{
\item{e1}{object of S4 class 'complex'}

\item{e2}{object of S4 class 'qstate'}
}
\value{
An object of S4 class 'qstate'
}
\description{
Multiplies a quantum gate by a global (phase) factor.
}
