% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqgate.R
\name{*,cqgate,qstate-method}
\alias{*,cqgate,qstate-method}
\title{times-cqgate-qstate}
\usage{
\S4method{*}{cqgate,qstate}(e1, e2)
}
\arguments{
\item{e1}{object of S4 class 'cqgate'}

\item{e2}{object of S4 class 'qstate'}
}
\value{
An object of S4 class 'qstate'
}
\description{
Applies a controlled single qubit gate to a quantum state.
}
