% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swapgate.R
\name{*,cswapgate,qstate-method}
\alias{*,cswapgate,qstate-method}
\title{times-cswapgate-qstate}
\usage{
\S4method{*}{cswapgate,qstate}(e1, e2)
}
\arguments{
\item{e1}{object of S4 class 'cswapgate'}

\item{e2}{object of S4 class 'qstate'}
}
\value{
An object of S4 class 'qstate'
}
\description{
Applies a CSWAP gate to a quantum state.
}
