% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{*,matrix,qstate-method}
\alias{*,matrix,qstate-method}
\title{times-matrix-qstate}
\usage{
\S4method{*}{matrix,qstate}(e1, e2)
}
\arguments{
\item{e1}{object of S4 class 'matrix'}

\item{e2}{object of S4 class 'qstate'}
}
\value{
An object of S4 class 'qstate'
}
\description{
Applies a single qubit gate to a quantum state.
}
