% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqgate.R
\name{*,sqgate,qstate-method}
\alias{*,sqgate,qstate-method}
\title{times-sqgate-qstate}
\usage{
\S4method{*}{sqgate,qstate}(e1, e2)
}
\arguments{
\item{e1}{object of S4 class 'sqgate'}

\item{e2}{object of S4 class 'qstate'}
}
\value{
An object of S4 class 'qstate'
}
\description{
Applies a single qubit gate to a quantum state.
}
