% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swapgate.R
\name{*,swapgate,qstate-method}
\alias{*,swapgate,qstate-method}
\title{times-swapgate-qstate}
\usage{
\S4method{*}{swapgate,qstate}(e1, e2)
}
\arguments{
\item{e1}{object of S4 class 'swapgate'}

\item{e2}{object of S4 class 'qstate'}
}
\value{
An object of S4 class 'qstate'
}
\description{
Applies a SWAP gate to a quantum state.
}
