% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqgate.R
\name{Y}
\alias{Y}
\title{The Y gate}
\usage{
Y(bit)
}
\arguments{
\item{bit}{integer. The bit to which to apply the gate}
}
\value{
An S4 class 'sqgate' object is returned
}
\description{
The Y gate
}
\examples{
x <- qstate(nbits=2)
z <- Y(1) * x
z

}
