% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqgate.R
\name{Z}
\alias{Z}
\title{The Z gate}
\usage{
Z(bit)
}
\arguments{
\item{bit}{integer. The bit to which to apply the gate}
}
\value{
An S4 class 'sqgate' object is returned
}
\description{
The Z gate
}
\examples{
x <- X(1) * qstate(nbits=2)
z <- Z(1) * x
z

}
